Rename reducers
